+++ lib/posix/options/Kconfig.clock_selection
@@ -1,0 +1,13 @@
+# Copyright (c) 2025 Tenstorrent AI ULC
+#
+# SPDX-License-Identifier: Apache-2.0
+
+config POSIX_CLOCK_SELECTION
+	bool "POSIX Clock Selection"
+	help
+	  Select 'y' here and Zephyr will provide an implementation of the POSIX_CLOCK_SELECTION Option
+	  Group, which includes the functions clock_nanosleep(), pthread_condattr_getclock(),
+	  pthread_condattr_setclock().
+
+	  For more information, please see
+	  https://pubs.opengroup.org/onlinepubs/9699919799/xrat/V4_subprofiles.html
